:art: More cleanups
