+++ circle.yml
@@ -8,6 +8,7 @@ dependencies:
     - npm --version
     - atom --version
     - rm -rf node_modules; npm install
+    - npm run compile
 
 test:
   override:
