+++ src/elements/message.js
@@ -1,13 +1,12 @@
 /* @flow */
 /** @jsx React.h */
 
-import React from 'preact'
+import { Component } from 'preact'
 import type { Message } from '../types'
 
-export default class MessageElement extends React.Component {
+export default class MessageElement extends Component {
   props: {
     message: Message,
-    reference: 'external' | 'internal',
     showProviderName: boolean,
   };
